fix: unify id types and implement setId for entities
Standardize the return type of getId and parameter type of setId to size_t
across all relevant classes (Pessoa, Equipamento, Plano, Professor,
PlanoPersonalizado, Avaliacao). Add missing setId declaration and
implementation for Avaliacao. This resolves type mismatches and ensures
interface consistency, addressing compilation errors and abstract class
instantiation issues.

+++ src/Equipamento.cpp
@@ -1,25 +1,28 @@
 #include "../header/Equipamento.h"
 
-Equipamento::Equipamento(std::string nome, int quantidade) : nome(std::move(nome)), quantidade(quantidade) {}
+Equipamento::Equipamento(std::string nome, int quantidade)
+    : nome(std::move(nome)), quantidade(quantidade) {}
 
 std::string Equipamento::getNome() const { return nome; }
 
 int Equipamento::getQuantidade() const { return quantidade; }
 
-int Equipamento::getId() const { return id; }
+size_t Equipamento::getId() const { return id; }
 
-void Equipamento::setId(int i) { id = i; }
+void Equipamento::setId(size_t i) { id = i; }
 
 void Equipamento::setQuantidade(int q) { quantidade = q; }
 
 void Equipamento::setNome(const std::string &n) { nome = n; }
 
 Equipamento Equipamento::operator+(const Equipamento &o) const {
-    if (nome != o.nome) return *this; // se nomes diferentes, retorna cópia simples
-    return Equipamento(nome, quantidade + o.quantidade);
+  if (nome != o.nome)
+    return *this; // se nomes diferentes, retorna cópia simples
+  return Equipamento(nome, quantidade + o.quantidade);
 }
 
 std::ostream &operator<<(std::ostream &os, const Equipamento &e) {
-    os << "ID:" << e.getId() << " " << e.getNome() << " (" << e.getQuantidade() << " repetições)";
-    return os;
+  os << "ID:" << e.getId() << " " << e.getNome() << " (" << e.getQuantidade()
+     << " repetições)";
+  return os;
 }
